Reorganized lstm related files to match Image Transformer file organization. Moved LSTM dataset wrangling to LSTM/lstm_dataset.py. Moved general data related functions to dataset_functions.py. Created basic util folder with function to check device type in utils.py. Deleted models.py and moved code to test.py in main folder to assist with easier module importing.
